feat: auth back up and running
